fix delete docs test refresh
