+++ tests/api_tests/test_add_documents.py
@@ -162,36 +162,6 @@ class TestAddDocuments(MarqoTestCase):
     def test_update_docs_updates_chunks(self):
         """TODO"""
 
-    # delete documents tests:
-
-    def test_delete_docs(self):
-        self.client.create_index(index_name=self.index_name_1)
-        self.client.index(self.index_name_1).add_documents([
-            {"abc": "wow camel", "_id": "123"},
-            {"abc": "camels are cool", "_id": "foo"}
-        ], non_tensor_fields=[])
-        res0 = self.client.index(self.index_name_1).search("wow camel")
-        print("res0res0")
-        pprint.pprint(res0)
-        assert res0['hits'][0]["_id"] == "123"
-        assert len(res0['hits']) == 2
-        self.client.index(self.index_name_1).delete_documents(["123"])
-        res1 = self.client.index(self.index_name_1).search("wow camel")
-        assert res1['hits'][0]["_id"] == "foo"
-        assert len(res1['hits']) == 1
-
-    def test_delete_docs_empty_ids(self):
-        self.client.create_index(index_name=self.index_name_1)
-        self.client.index(self.index_name_1).add_documents([{"abc": "efg", "_id": "123"}], non_tensor_fields=[])
-        try:
-            self.client.index(self.index_name_1).delete_documents([])
-            raise AssertionError
-        except MarqoWebError as e:
-            assert "can't be empty" in str(e) or "value_error.missing" in str (e)
-        res = self.client.index(self.index_name_1).get_document("123")
-        print(res)
-        assert "abc" in res
-
     # get documents tests :
 
     def test_get_document(self):
